Add files via upload
I am thrilled to present my completed project, a redesign of Apple's homepage. This project was undertaken as an exercise in honing my web design skills and exploring creative approaches to user interface and experience. The goal was to provide a fresh and modern look while maintaining the essence of Apple's brand identity.